Fix voting bug
Now when the user upvotes and downvotes a post, the behaviour is as
expected.

The backend change to views.py was to send the whole queryset each time
instead of filtering posts by matching user_id . When filtering, only
the votes for the current user were being shown, and not the total votes
for a post.

In the frontend, many changes were made and the code has been
simplified.