Improve help section search
Adding a basic fuzzy search makes it easier for users to search the help
section.

It has been implemented in the same way as in views/EventsViews.vue.
Ideally we would put the searching function in a separate javascript
file and that way we wouldn't have to duplicate the code, but this way
makes it easier to integrate with the final version.