Add manual tagging
Prefixing a word with @ will turn it into a tag. The tag just makes the
text blue, it isn't a link. There is no check whether the word actually
matches a user. It should be secure against web attacks as raw html
isn't evaluated.

Also removed the unused function `dbg`.